refactor: split model_factory into backend/ modules; unify vLLM engine classes
