+++ nemo/collections/speechlm2/inference/model_wrappers/backend/__init__.py
@@ -1,0 +1,23 @@
+# Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from nemo.collections.speechlm2.inference.model_wrappers.backend.interface import ModelInterface
+from nemo.collections.speechlm2.inference.model_wrappers.backend.pytorch.model import PyTorchLLM
+from nemo.collections.speechlm2.inference.model_wrappers.backend.pytorch.eartts import PyTorchEarTTS, TTSGenerationResult
+
+try:
+    from nemo.collections.speechlm2.inference.model_wrappers.backend.vllm.llm import VLLMLLM
+    from nemo.collections.speechlm2.inference.model_wrappers.backend.vllm.eartts import VLLMEarTTS
+except ImportError:
+    pass
