--- nemo/collections/speechlm2/inference/model_wrappers/model_factory.py
@@ -1,1170 +1,0 @@
-# Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Model Interface for S2S Inference
-
-This module provides an abstract interface for model inference engines,
-allowing seamless swapping between different implementations (e.g., native PyTorch, vLLM)
-without modifying the inference code.
-
-Usage Example:
-    from nemo.collections.speechlm2.inference.model_wrappers.model_factory import create_model
-
-    # Create interface (automatically wraps existing model)
-    model_interface = create_model(
-        model=your_model,
-        engine_type="native"  # or "vllm"
-    )
-
-    # Use the interface exactly as you would use self.model()
-    ans = model_interface(input_embeds, cache=cache)
-"""
-
-from abc import ABC, abstractmethod
-from typing import Any
-import math
-import os
-import torch
-from dataclasses import dataclass, fields
-
-from nemo.utils import logging
-
-class ModelInterface(ABC):
-    """
-    Base interface for model inference engines with shared sampling utilities.
-
-    This interface defines the contract that all model implementations must follow,
-    ensuring consistent behavior across different engine types. It also provides
-    concrete implementations of sampling methods (top-p, repetition penalty) that
-    can be shared across all engines.
-    """
-
-    def __init__(
-        self,
-        special_token_ids: set[int] | None = None,
-        top_p: float = 1.0,
-        repetition_penalty: float = 1.0,
-        temperature: float = 1.0,
-    ):
-        """
-        Initialize base interface with sampling parameters.
-
-        Args:
-            special_token_ids: Set of special token IDs (pad, eos, bos) that should bypass sampling.
-                               These tokens will use greedy decoding and won't be penalized.
-            top_p: Top-p (nucleus) sampling threshold. 1.0 disables it (greedy). Default: 1.0
-            repetition_penalty: Penalty for repeated tokens. 1.0 disables it. Default: 1.0
-            temperature: Temperature for sampling. 1.0 = no change, <1.0 = sharper, >1.0 = flatter.
-                        0.0 = greedy (argmax). Default: 1.0
-        """
-        if not math.isfinite(temperature):
-            raise ValueError(f"temperature must be finite, got {temperature}")
-        if temperature < 0.0:
-            raise ValueError(f"temperature must be >= 0.0, got {temperature}")
-
-        self.special_token_ids = special_token_ids if special_token_ids is not None else set()
-        self.top_p = top_p
-        self.repetition_penalty = repetition_penalty
-        self.temperature = temperature
-
-        # Pre-built tensor for special-token filtering in repetition penalty.
-        # Lazily moved to the right device on first use (see _sample_text_token).
-        self._special_ids_tensor: torch.Tensor | None = (
-            torch.tensor(sorted(self.special_token_ids), dtype=torch.long)
-            if self.special_token_ids else None
-        )
-
-    def _sample_text_token(
-        self,
-        logits: torch.Tensor,
-        generated_tokens: torch.Tensor,
-        current_step: int,
-        sampling_params: dict[str, float] | None = None,
-    ) -> torch.Tensor:
-        """
-        Sample text tokens with optional top-p sampling and repetition penalty.
-        Special tokens (pad, eos, bos) bypass sampling - if they have highest probability, return them directly.
-
-        Args:
-            logits: Logits tensor of shape (B, V) for vocabulary V.
-            generated_tokens: Previously generated tokens of shape (B, T).
-            current_step: Current decoding step (used to slice generated_tokens).
-            sampling_params: Optional per-request overrides for ``top_p``,
-                ``temperature``, and ``repetition_penalty``.  Missing keys
-                fall back to ``self.*`` (the pipeline-level defaults).
-
-        Returns:
-            Sampled token ids of shape (B,).
-        """
-        top_p = sampling_params.get("top_p", self.top_p) if sampling_params else self.top_p
-        temperature = sampling_params.get("temperature", self.temperature) if sampling_params else self.temperature
-        rep_penalty = sampling_params.get("repetition_penalty", self.repetition_penalty) if sampling_params else self.repetition_penalty
-
-        B, V = logits.shape
-        device = logits.device
-
-        # First check greedy tokens (on original logits)
-        greedy_tokens = logits.argmax(dim=-1)  # (B,)
-
-        # If no sampling needed (all disabled), return greedy
-        if top_p >= 1.0 and rep_penalty == 1.0 and (temperature == 1.0 or temperature == 0.0):
-            return greedy_tokens
-
-        # temperature=0 means greedy
-        if temperature == 0.0:
-            return greedy_tokens
-
-        # For each batch, if greedy is special token, use greedy; otherwise sample
-        sampled_tokens = greedy_tokens.clone()
-
-        # Ensure cached special-token tensor is on the right device (once).
-        if self._special_ids_tensor is not None and self._special_ids_tensor.device != device:
-            self._special_ids_tensor = self._special_ids_tensor.to(device)
-
-        for b in range(B):
-            # If greedy token is a special token, keep it (no sampling)
-            if greedy_tokens[b].item() in self.special_token_ids:
-                continue
-
-            # Not a special token - apply repetition penalty and sampling
-            batch_logits = logits[b].clone()  # (V,)
-
-            # Apply repetition penalty (vectorized, no Python loop)
-            if rep_penalty != 1.0 and current_step > 0:
-                unique_prev = generated_tokens[b, :current_step].unique()
-                # Exclude special tokens from penalty
-                if self._special_ids_tensor is not None:
-                    ids_t = self._special_ids_tensor
-                    if ids_t.device != unique_prev.device:
-                        ids_t = ids_t.to(unique_prev.device)
-                    unique_prev = unique_prev[~torch.isin(unique_prev, ids_t)]
-
-                if unique_prev.numel() > 0:
-                    if unique_prev.device != batch_logits.device:
-                        unique_prev = unique_prev.to(batch_logits.device)
-                    prev_logits = batch_logits[unique_prev]
-                    # Positive logits are divided, negative logits are multiplied
-                    # (same as the standard repetition_penalty convention)
-                    batch_logits[unique_prev] = torch.where(
-                        prev_logits > 0,
-                        prev_logits / rep_penalty,
-                        prev_logits * rep_penalty,
-                    )
-
-            # Apply temperature scaling
-            if temperature != 1.0:
-                batch_logits = batch_logits / temperature
-
-            # Fall back to greedy if logits are non-finite before top-p
-            # (top-p intentionally introduces -inf, so check must happen first)
-            if not torch.isfinite(batch_logits).all():
-                logging.warning(
-                    f"_sample_text_token: logits contain NaN or inf at step {current_step}, batch {b}: "
-                    f"nan={batch_logits.isnan().sum().item()}, "
-                    f"inf={batch_logits.isinf().sum().item()}, "
-                    f"min={batch_logits[~batch_logits.isnan()].min().item() if not batch_logits.isnan().all() else 'all_nan'}, "
-                    f"max={batch_logits[~batch_logits.isnan()].max().item() if not batch_logits.isnan().all() else 'all_nan'}"
-                )
-                sampled_tokens[b] = greedy_tokens[b]
-                continue
-
-            # Apply top-p (nucleus) sampling
-            if top_p < 1.0:
-                sorted_logits, sorted_indices = torch.sort(batch_logits, descending=True)
-                sorted_probs = torch.softmax(sorted_logits, dim=-1)
-                cumulative_probs = torch.cumsum(sorted_probs, dim=-1)
-
-                # Remove tokens with cumulative prob > top_p, keeping at least one
-                sorted_indices_to_remove = cumulative_probs > top_p
-                # Shift to keep the first token that exceeds threshold
-                sorted_indices_to_remove[1:] = sorted_indices_to_remove[:-1].clone()
-                sorted_indices_to_remove[0] = False
-
-                # Set to -inf
-                indices_to_remove = sorted_indices[sorted_indices_to_remove]
-                batch_logits[indices_to_remove] = float('-inf')
-
-            probs = torch.softmax(batch_logits, dim=-1)
-            sampled_tokens[b] = torch.multinomial(probs, num_samples=1).item()
-
-        return sampled_tokens
-
-    @abstractmethod
-    def __call__(
-        self,
-        input_embeds: torch.Tensor,
-        cache: Any | None = None,
-        **kwargs
-    ) -> dict[str, Any]:
-        """
-        Perform model inference.
-
-        Args:
-            input_embeds: Input embeddings tensor of shape [batch, seq_len, hidden_dim]
-            cache: Optional cache object (e.g., DynamicCache for transformers)
-            **kwargs: Additional model-specific arguments
-
-        Returns:
-            Dictionary containing:
-                - 'text_logits': Logits for text generation [batch, seq_len, vocab_size]
-                - 'cache': Updated cache object (if cache was provided)
-                - Additional model-specific outputs
-        """
-        pass
-
-    @abstractmethod
-    def to(self, device_or_dtype: torch.device | torch.dtype) -> 'ModelInterface':
-        """Move model to specified device or convert to specified dtype."""
-        pass
-
-    @abstractmethod
-    def eval(self) -> 'ModelInterface':
-        """Set model to evaluation mode."""
-        pass
-
-    @property
-    @abstractmethod
-    def device(self) -> torch.device:
-        """Get the device of the model."""
-        pass
-
-
-class VllmLLMModel(ModelInterface):
-    """
-    vLLM-based model interface using LLMStreamingEngine.
-
-
-    This wraps the LLMStreamingEngine to provide async streaming inference
-    while conforming to the ModelInterface contract. Supports multiple concurrent
-    requests sharing a single engine instance.
-
-    model = VllmLLMModel(...)
-
-    async def process_stream(embeds, stream_id):
-        # Use the async engine directly
-        result = await model._async_inference(embeds, f"stream_{stream_id}", seq_len)
-        return result
-
-    # Run multiple streams concurrently in same event loop
-    async def main():
-        results = await asyncio.gather(
-            process_stream(embeds1, 1),
-            process_stream(embeds2, 2),
-            process_stream(embeds3, 3)
-        )
-
-    asyncio.run(main())
-    """
-
-    def __init__(
-        self,
-        model_path: str,
-        max_model_len: int = 1024,
-        gpu_memory_utilization: float = 0.8,
-        trust_remote_code: bool = True,
-        dtype: str = "bfloat16",
-        engine_path: str | None = None,
-        pretrained_llm: str | None = None,
-        special_token_ids: set[int] | None = None,
-        top_p: float = 1.0,
-        repetition_penalty: float = 1.0,
-        temperature: float = 1.0,
-        model_type: str = "llm",
-        **sampling_kwargs
-    ):
-        """
-        Initialize vLLM model interface with LLMStreamingEngine.
-
-        Args:
-            model_path: Path to the vLLM-compatible model checkpoint
-            max_model_len: Maximum sequence length
-            gpu_memory_utilization: GPU memory utilization ratio (0.0-1.0)
-            trust_remote_code: Whether to trust remote code in model
-            dtype: Data type for embeddings (e.g., "bfloat16", "float16")
-            engine_path: Optional path to pre-converted vLLM model
-            pretrained_llm: Optional path to pretrained LLM for conversion
-            special_token_ids: Set of special token IDs (for potential post-processing)
-            top_p: Top-p sampling (currently vLLM uses greedy decoding)
-            repetition_penalty: Repetition penalty (currently not used by vLLM engine)
-            temperature: Temperature for sampling. Applied in _sample_text_token, not in vLLM engine.
-            model_type: Type of model for vLLM engine ("llm", "chatglm", etc.)
-            **sampling_kwargs: Additional sampling parameters passed to vLLM engine.
-                               By default, vLLM uses greedy decoding (temperature=0)
-        """
-        # Initialize base class with sampling parameters
-        super().__init__(
-            special_token_ids=special_token_ids,
-            top_p=top_p,
-            repetition_penalty=repetition_penalty,
-            temperature=temperature,
-        )
-
-        import asyncio
-        from nemo.collections.speechlm2.inference.vllm.streaming_llm_engine import LLMStreamingEngine
-
-        self.model_path = model_path
-        self.pretrained_llm = pretrained_llm
-        self._dtype = dtype
-        self._device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        # Force greedy decoding in vLLM by setting temperature=0 if not specified
-        if 'temperature' not in sampling_kwargs:
-            sampling_kwargs['temperature'] = 0.0
-
-        if engine_path is None:
-            # convert model to vLLM format if needed
-            dir_name = os.path.basename(os.path.normpath(model_path))
-            engine_path = "/tmp/" + dir_name + f"_vllm_converted_{model_type}"
-            if os.path.exists(engine_path):
-                logging.info(f"Found existing vLLM converted model at {engine_path}")
-            else:
-                self._convert_ckpt(
-                    save_path=engine_path
-                )
-
-        from nemo.collections.speechlm2.inference.vllm.streaming_llm_engine import create_engine
-        # Initialize the streaming engine
-        self.engine = create_engine(
-            engine_type=model_type,
-            model_path=engine_path,
-            max_model_len=max_model_len,
-            gpu_memory_utilization=gpu_memory_utilization,
-            trust_remote_code=trust_remote_code,
-            dtype=dtype,
-            **sampling_kwargs
-        )
-        # Track request counter
-        self._request_counter = 0
-
-        # Get or create event loop
-        try:
-            self._loop = asyncio.get_event_loop()
-        except RuntimeError:
-            self._loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(self._loop)
-
-        # Initialize engine immediately to avoid first-call latency
-        logging.info("Initializing vLLM engine (this may take a moment)...")
-        self._loop.run_until_complete(self.engine.initialize())
-
-        if self.engine.engine.tokenizer is not None and not self.special_token_ids:
-            self.special_token_ids = self._get_special_token_ids_from_vllm_tokenizer(self.engine.engine.tokenizer)
-
-        logging.debug(f"Special token IDs: {self.special_token_ids}")
-        logging.info("vLLM engine ready!")
-
-    @staticmethod
-    def _get_special_token_ids_from_vllm_tokenizer(tokenizer) -> set[int]:
-        """
-        Extract special token IDs from a vLLM tokenizer.
-        Looks for: '<s>' (bos), '</s>' (eos), '<SPECIAL_12>' (pad).
-
-        Args:
-            tokenizer: A vLLM CachedTokenizer instance.
-
-        Returns:
-            Set of special token IDs.
-        """
-        special_ids = set()
-        for token in ('<s>', '</s>', '<SPECIAL_12>'):
-            try:
-                tid = tokenizer.convert_tokens_to_ids(token)
-                if isinstance(tid, int):
-                    special_ids.add(tid)
-            except Exception:
-                pass
-        return special_ids
-
-    def _convert_ckpt(self, save_path: str):
-        """Convert existing checkpoint to vLLM format and save."""
-        from nemo.collections.speechlm2.inference.vllm.scripts.convert_nemotronllm_checkpoint import convert_nemo_to_hf_format
-
-        convert_nemo_to_hf_format(
-            checkpoint_path=self.model_path,
-            output_dir=save_path,
-            pretrained_llm=self.pretrained_llm,
-            dtype=self._dtype
-        )
-        logging.info(f"Converted model saved to {save_path}")
-
-    def _generate_request_id(self) -> str:
-        """Generate a unique request ID."""
-        self._request_counter += 1
-        return f"vllm_request_{self._request_counter}"
-
-    def __call__(
-        self,
-        input_embeds: torch.Tensor,
-        request_id: str | None = "request_id_1",
-        **kwargs
-    ) -> dict[str, Any]:
-        """
-        Perform inference using vLLM streaming engine.
-
-        Args:
-            inputs:
-            cache: Optional cache object (currently not used for streaming)
-            generated_tokens: Optional tensor of generated tokens
-            current_step: Current decoding step
-            request_id: Unique request identifier for this generation
-            **kwargs: Additional model-specific arguments
-
-        Returns:
-            Dictionary containing:
-                - predicted_token: Last generated text token
-                - asr_predicted_token: Last generated ASR token
-                - cache: None (vLLM manages cache internally)
-                - is_finished: Whether generation is complete
-                - request_id: The request identifier
-        """
-        # Run async inference
-        result = self._loop.run_until_complete(
-            self._async_inference(input_embeds, request_id, **kwargs)
-        )
-        return result
-
-    async def _async_inference(
-        self,
-        inputs: torch.Tensor | list[torch.Tensor],
-        request_id: str,
-        **kwargs
-    ) -> dict[str, Any]:
-        """
-        Async inference using the streaming engine.
-
-        Args:
-            input_embeds: Input embeddings [batch, seq_len, hidden_dim]
-            request_id: Unique request identifier
-            seq_len: Number of decoding steps to perform
-
-        Returns:
-            Dictionary with text_logits and other outputs
-        """
-        # Check request status and restart if needed
-        from nemo.collections.speechlm2.inference.vllm.streaming_llm_engine import StreamStatus
-
-        if request_id not in self.engine.requests:
-            await self.engine.start_generation(request_id=request_id)
-        else:
-            # Check if request is finished and needs restart
-            request_state = self.engine.requests[request_id]
-            if request_state.status in (StreamStatus.FINISHED, StreamStatus.ABORTED):
-                logging.warning(
-                    f"Request {request_id} was {request_state.status.value}. "
-                    f"Generated {len(request_state.generated_tokens)} tokens before stopping. "
-                    "Cleaning up and restarting..."
-                )
-                # Try to abort cleanly first
-                try:
-                    await self.engine.abort_generation(request_id)
-                except Exception:
-                    pass
-                # Start fresh
-                await self.engine.start_generation(request_id=request_id)
-
-        # Process embeddings to generate tokens
-        return await self._process_inputs_to_outputs(inputs, request_id, **kwargs)
-
-    async def _process_inputs_to_outputs(
-        self,
-        input_embeds: torch.Tensor,
-        request_id: str,
-        decode_steps: int = 1,
-        prompt_token_ids: list | None = None,
-        generated_tokens: torch.Tensor | None = None,
-        current_step: int = 0,
-        sampling_params: dict[str, float] | None = None,
-    ) -> dict[str, Any]:
-        """
-        Process embeddings sequentially to generate text and ASR tokens.
-
-        Args:
-            input_embeds: Input embeddings [batch, seq_len, hidden_dim]
-            request_id: Request identifier
-            decode_steps: Number of decoding steps to perform; decode steps = 0 means prefill
-            prompt_token_ids: Optional list of prompt token IDs for prefill
-            generated_tokens: Previously generated tokens [batch, num_generated].
-                             Required for repetition_penalty. If None, creates empty tensor.
-            current_step: Current decoding step. Used for repetition penalty.
-            sampling_params: Optional per-request overrides for sampling.
-        """
-
-        if decode_steps == 0:
-            # prefill only, no token generation
-            input_embeds = input_embeds.flatten(0, 1)  # [seq_len, hidden_dim]
-            result = await self.engine.generate_next_token([input_embeds],
-                                                            prompt_token_ids,
-                                                            request_id=request_id)
-            return True if result is not None else False
-
-        # Process each embedding in sequence
-        text_token_ids = []
-        asr_token_ids = []
-        result = None
-        for i in range(decode_steps):
-            # Extract single embedding [1, hidden_dim]
-            single_embed = input_embeds[:, i:i+1, :].squeeze(1)  # [batch, hidden_dim]
-
-            # Generate next token
-            result = await self.engine.generate_next_token([single_embed], request_id=request_id)
-            if result is None:
-                # No token generated (finished or error)
-                break
-
-            text_token_ids.append(result.token_id)
-            asr_token_ids.append(result.custom_outputs["asr_tokens"])  # Assuming custom_outputs contains asr tokens
-
-            if result.is_finished:
-                break
-
-        assert len(text_token_ids) <= decode_steps, "Generated more tokens than input embeddings"
-        # Handle case when no tokens were generated
-        is_finished = False
-        if text_token_ids:
-            is_finished = len(text_token_ids) < decode_steps or (result and result.is_finished)
-
-        text_logits = result.custom_outputs["text_logits"] if result else None
-
-        predicted_token = text_token_ids[-1]
-        eff_top_p = sampling_params.get("top_p", self.top_p) if sampling_params else self.top_p
-        eff_temp = sampling_params.get("temperature", self.temperature) if sampling_params else self.temperature
-        eff_rep = sampling_params.get("repetition_penalty", self.repetition_penalty) if sampling_params else self.repetition_penalty
-        if eff_top_p < 1.0 or eff_rep != 1.0 or (eff_temp != 1.0 and eff_temp != 0.0):
-            # Use provided generated_tokens or create empty tensor
-            batch_size = text_logits.shape[0]
-            if generated_tokens is None:
-                gen_tokens = torch.empty(batch_size, 0, device=text_logits.device, dtype=torch.long)
-            else:
-                gen_tokens = generated_tokens
-
-            # Apply sampling with top-p and repetition penalty
-            predicted_token = self._sample_text_token(
-                logits=text_logits,
-                generated_tokens=gen_tokens,
-                current_step=current_step,
-                sampling_params=sampling_params,
-            )
-
-        ans = {
-            "predicted_token": predicted_token,
-            "asr_predicted_token": asr_token_ids[-1],
-            "cache": None,  # vLLM manages cache internally
-            "is_finished": is_finished,
-            "request_id": request_id
-        }
-        if result and result.custom_outputs and "function_tokens" in result.custom_outputs:
-            ans["function_predicted_token"] = result.custom_outputs["function_tokens"]
-        return ans
-
-
-    def to(self, device_or_dtype: torch.device | torch.dtype) -> 'VllmLLMModel':
-        """
-        Move model to specified device or convert to specified dtype.
-
-        Note: vLLM manages device placement internally, this is for compatibility.
-        """
-        if isinstance(device_or_dtype, torch.device):
-            self._device = device_or_dtype
-        elif isinstance(device_or_dtype, torch.dtype):
-            # dtype conversion not directly supported, update config
-            pass
-        return self
-
-    def eval(self) -> 'VllmLLMModel':
-        """Set model to evaluation mode (vLLM is always in eval mode)."""
-        return self
-
-    @property
-    def device(self) -> torch.device:
-        """Get the device of the model."""
-        return self._device
-
-    def abort_request(self, request_id: str) -> bool:
-        """
-        Abort a specific generation request.
-
-        Args:
-            request_id: Request identifier to abort
-
-        Returns:
-            bool: True if abort was successful
-        """
-        return self._loop.run_until_complete(
-            self.engine.abort_generation(request_id)
-        )
-
-    def restart_request(self, request_id: str) -> bool:
-        """
-        Restart a finished or aborted generation request.
-
-        Args:
-            request_id: Request identifier to restart
-
-        Returns:
-            bool: True if restart was successful
-        """
-        # First abort if active
-        if request_id in self.engine.requests:
-            self.abort_request(request_id)
-
-        # Start new generation
-        return self._loop.run_until_complete(
-            self.engine.start_generation(request_id=request_id)
-        )
-
-    def get_request_status(self, request_id: str | None = None) -> dict[str, Any]:
-        """
-        Get status of a specific request or all requests.
-
-        Args:
-            request_id: Optional request ID. If None, returns all requests.
-
-        Returns:
-            Status dictionary
-        """
-        return self.engine.get_status(request_id)
-
-    def shutdown(self):
-        """Shutdown the vLLM engine and cleanup resources."""
-        self._loop.run_until_complete(self.engine.shutdown())
-
-    def __del__(self):
-        """Cleanup on deletion."""
-        try:
-            self.shutdown()
-        except Exception:
-            pass
-
-@dataclass
-class TTSGenerationResult:
-    codes: torch.Tensor  # Generated acoustic tokens
-    past_key_values: Any  # Updated cache (if applicable)
-
-    def __getitem__(self, item: str | int):
-        """Allows for accessing attributes by key or index."""
-        if isinstance(item, str):
-            return getattr(self, item)
-        else:
-            # Access fields in the order they are defined in the dataclass
-            return getattr(self, fields(self)[item].name)
-
-
-class VllmEARTTSModel(VllmLLMModel):
-    """
-    vLLM-based model interface specialized for EARTTS models.
-
-    Inherits from VllmLLMModel and sets EARTTS-specific configurations.
-    """
-
-    def __init__(self, **kwargs):
-        """
-        Initialize vLLM EARTTS model interface.
-
-        Args:
-            **kwargs: Arguments passed to the VllmLLMModel constructor
-        """
-        super().__init__(**kwargs)
-        self._speaker_latent_dim = None
-        logging.info("VllmEARTTSModel initialized with EARTTS-specific settings.")
-
-    def _convert_ckpt(self, save_path: str):
-        """Convert EARTTS checkpoint to vLLM format."""
-        from nemo.collections.speechlm2.inference.vllm.scripts.convert_eartts_checkpoint import convert
-        ckpt_dir = os.path.normpath(self.model_path)
-        config_file = os.path.join(ckpt_dir, "config.json")
-        model_ckpt = os.path.join(ckpt_dir, "model.safetensors")
-        convert(save_path, config_file, model_ckpt)
-
-    def __call__(
-        self,
-        inputs: dict[str, torch.Tensor] | None = None,
-        request_id: str | None = None,
-        prompt_token_ids: list | None = None,
-        **kwargs
-    ) -> TTSGenerationResult:
-        """
-        Perform TTS inference using vLLM streaming engine.
-
-        Supports two calling conventions:
-        1. model(inputs_dict, request_id="id")  - pass dict as first positional arg
-        2. model(**inputs_dict)  - unpack dict as keyword arguments
-
-        Args:
-            inputs: Optional dict of model inputs (if None, uses **kwargs)
-            request_id: Optional request identifier
-            **kwargs: Model inputs as keyword arguments (used if inputs is None):
-                - code: prev_audio_tokens
-                - context_hidden_state: context_hidden_state (must be None)
-                - subword_ids: current_subword_id
-                - subword_mask: current_subword_mask
-                - past_key_values: past_key_values
-                - use_cache: True
-                - guidance_enabled: guidance_enabled
-                - generation_config: generation_config
-                - ignore_eos_flag_stop: ignore_eos_flag_stop
-
-        Returns:
-            TTSGenerationResult containing generated acoustic tokens and cache
-        """
-        # Handle both calling conventions
-        if inputs is not None:
-            # Called as model(inputs_dict, request_id="id")
-            input_dict = inputs
-        else:
-            # Called as model(**inputs_dict)
-            # Extract request_id from kwargs if present
-            if request_id is None:
-                request_id = kwargs.pop('request_id', None)
-            input_dict = kwargs
-
-        # Use default request_id if still None
-        if request_id is None:
-            request_id = 'tts_request_id_1'
-
-        # Run async inference
-        result = self._loop.run_until_complete(
-            self._async_inference(input_dict, request_id, prompt_token_ids=prompt_token_ids)
-        )
-
-        return result
-
-    async def _process_inputs_to_outputs(
-        self,
-        inputs: dict[str, torch.Tensor],
-        request_id: str,
-        prompt_token_ids: list | None = None,
-    ) -> dict[str, Any]:
-        """
-        Process embeddings sequentially to generate text and ASR tokens.
-
-        Args:
-        inputs = {
-            "code": prev_audio_tokens,
-            "context_hidden_state": context_hidden_state,
-            "subword_ids": current_subword_id,
-            "subword_mask": current_subword_mask,
-            "past_key_values": past_key_values,
-            "use_cache": True,
-            "guidance_enabled": guidance_enabled,
-            "generation_config": generation_config,
-            "ignore_eos_flag_stop": ignore_eos_flag_stop,
-            }
-        Returns:
-            step_acoustic_tokens: Generated acoustic tokens for the current step
-            cache: None (vLLM manages cache internally)
-        """
-
-        assert inputs["context_hidden_state"] is None, "EARTTS vllm model does not support context_hidden_state input"
-
-        codes = inputs["code"].squeeze(0)  # T x 31
-        if codes.shape[0] > 1:
-            # in prefill stage, we needto shift acoustic tokens for vllm,
-            # replicating the NeMo logic from here:
-            # https://github.com/erastorgueva-nv/NeMo/blob/duplex-realtime-inference/nemo/collections/speechlm2/modules/ear_tts_model.py#L1357
-            codes = torch.nn.functional.pad(codes[:-1], [0, 0, 1, 0])
-        input_tensors = [
-            codes,
-            inputs["subword_ids"].squeeze(0),
-            inputs["subword_mask"].squeeze(0),
-        ]
-
-        if "non_prompt_mask" in inputs:
-            # Apply edge detection to match native model's BOS placement logic:
-            # BOS should only be applied at the FIRST position where non_prompt_mask is True
-            non_prompt_mask = inputs["non_prompt_mask"].squeeze(0)  # T
-            # Compute edge: positions where mask is True AND previous position is False
-            padded_prev = torch.nn.functional.pad(non_prompt_mask[:-1], [1, 0], value=False)
-            bos_mask = (non_prompt_mask & (~padded_prev)).to(dtype=getattr(torch, self._dtype))
-            input_tensors.append(bos_mask)
-
-
-        else:
-            current_subword_id = input_tensors[1]
-            # Use a tiny epsilon instead of exact 0 so the vLLM model's
-            # (bos_mask == 0) check is False during decoding.  This prevents
-            # use_audio_prompt_frozen_projection from incorrectly applying the
-            # speaker-prompt projection to every decoding step.  The epsilon is
-            # small enough that bos_mask * bos_emb remains negligible.
-            bos_mask = torch.full_like(current_subword_id, 1e-20, dtype=getattr(torch, self._dtype))
-            input_tensors.append(bos_mask)
-
-        # Pass speaker_latent: the pre-extracted speaker embedding.
-        # During prefill with speaker_name: audio_prompt_lantent is [1, T, hidden_size]
-        # During decode or speaker_reference: pass zeros so the model falls back
-        # to computing the latent from acoustic tokens.
-        if "audio_prompt_lantent" in inputs and inputs["audio_prompt_lantent"] is not None:
-            speaker_latent = inputs["audio_prompt_lantent"].squeeze(0)  # T x hidden_size
-            self._speaker_latent_dim = speaker_latent.shape[-1]
-            input_tensors.append(speaker_latent.to(dtype=getattr(torch, self._dtype)))
-        else:
-            if self._speaker_latent_dim is None:
-                # Read hidden_size from the converted model config
-                import json as _json
-                dir_name = os.path.basename(os.path.normpath(self.model_path))
-                converted_config_path = os.path.join("/tmp", dir_name + "_vllm_converted_eartts", "config.json")
-                if os.path.exists(converted_config_path):
-                    with open(converted_config_path) as _f:
-                        self._speaker_latent_dim = _json.load(_f)["hidden_size"]
-                else:
-                    raise RuntimeError(
-                        f"Cannot determine speaker_latent_dim: converted config not found at {converted_config_path}. "
-                        "Run a prefill with audio_prompt_lantent first, or ensure the converted checkpoint exists."
-                    )
-            num_tokens = codes.shape[0]
-            speaker_latent = torch.zeros(num_tokens, self._speaker_latent_dim, dtype=getattr(torch, self._dtype))
-            input_tensors.append(speaker_latent)
-
-        result = await self.engine.generate_next_token(input_tensors, prompt_token_ids=prompt_token_ids, request_id=request_id)
-        acoustic_tokens = result.custom_outputs["acoustic_tokens"]  # T x 31
-        step_acoustic_tokens = acoustic_tokens[-1:]  # 1 x 31
-        return TTSGenerationResult(
-            codes=step_acoustic_tokens.unsqueeze(0).cuda(),  # Add batch dim back: 1 x 1 x 31
-            past_key_values=None  # vLLM manages cache internally
-        )
-
-class NativeModel(ModelInterface):
-    """
-    Native PyTorch model interface.
-
-    This wraps the existing DuplexS2SExternalSpeechDecoderModel to conform
-    to the ModelInterface contract. Supports top-k, top-p sampling and repetition penalty.
-    """
-
-    def __init__(
-        self,
-        model,
-        special_token_ids: set[int] | None = None,
-        top_p: float = 1.0,
-        repetition_penalty: float = 1.0,
-        temperature: float = 1.0,
-    ):
-        """
-        Initialize with an existing model.
-
-        Args:
-            model: The DuplexS2SExternalSpeechDecoderModel instance
-            special_token_ids: Set of special token IDs (pad, eos, bos) that should bypass sampling.
-                               These tokens will use greedy decoding and won't be penalized.
-                               If None, will try to extract from model.tokenizer for tokens:
-                               '<s>' (bos), '</s>' (eos), '<SPECIAL_12>' (pad).
-                               You can also manually provide: {tokenizer.pad_token_id, tokenizer.eos_token_id, tokenizer.bos_token_id}
-            top_p: Top-p (nucleus) sampling threshold. 1.0 disables it (greedy). Default: 1.0
-            repetition_penalty: Penalty for repeated tokens. 1.0 disables it. Default: 1.0
-                               Recommended value when enabling: 1.2
-            temperature: Temperature for sampling. 1.0 = no change, <1.0 = sharper, >1.0 = flatter.
-                        0.0 = greedy (argmax). Default: 1.0
-        """
-        # Default special token IDs: bos=1, eos=2, pad=12
-        DEFAULT_SPECIAL_TOKEN_IDS = {1, 2, 12}
-
-        # Try to extract special token IDs from model if not provided
-        if special_token_ids is None:
-            special_token_ids = self._extract_special_token_ids_from_nemo(model)
-        # Fallback to default if extraction failed
-        if not special_token_ids:
-            special_token_ids = DEFAULT_SPECIAL_TOKEN_IDS
-        # Initialize base class with sampling parameters
-        super().__init__(
-            special_token_ids=special_token_ids,
-            top_p=top_p,
-            repetition_penalty=repetition_penalty,
-            temperature=temperature,
-        )
-
-        self.model = model
-
-        logging.debug(f"Special token IDs: {self.special_token_ids}")
-
-        # Validate: if sampling is enabled, special_token_ids should be set
-        sampling_active = top_p < 1.0 or repetition_penalty != 1.0 or (temperature != 1.0 and temperature != 0.0)
-        if sampling_active and not self.special_token_ids:
-            import warnings
-            warnings.warn(
-                "Sampling is enabled but special_token_ids is empty. "
-                "Could not auto-extract from model.tokenizer. "
-                "Please provide special_token_ids manually to ensure special tokens use greedy decoding. "
-                "Otherwise, EOS tokens may be randomly sampled and generation may not stop properly!"
-            )
-
-    def __call__(
-        self,
-        input_embeds: torch.Tensor,
-        cache: Any | None = None,
-        cache_position: torch.Tensor | None = None,
-        generated_tokens: torch.Tensor | None = None,
-        current_step: int = 0,
-        return_logits: bool = False,
-        sampling_params: dict[str, float] | None = None,
-        **kwargs
-    ) -> dict[str, Any]:
-        """
-        Perform inference using the native model.
-
-        Args:
-            input_embeds: Input embeddings [batch, seq_len, hidden_dim]
-            cache: Optional DynamicCache or HybridMambaAttentionDynamicCache
-            cache_position: Optional position tensor for Nemotron models
-            generated_tokens: Previously generated tokens [batch, num_generated].
-                             Required for repetition_penalty. If None, creates empty tensor.
-            current_step: Current decoding step. Used for repetition penalty.
-            sampling_params: Optional per-request overrides for sampling
-                (top_p, temperature, repetition_penalty).
-            **kwargs: Additional arguments passed to the model
-
-        Returns:
-            Dictionary with 'predicted_token', 'asr_predicted_token', and 'cache'
-        """
-        result = self.model.stt_model(input_embeds, cache=cache, cache_position=cache_position, **kwargs)
-
-        # Ensure consistent return format
-        if not isinstance(result, dict):
-            raise TypeError(f"Model returned {type(result)}, expected dict")
-
-        if 'text_logits' not in result:
-            raise KeyError("Model output must contain 'text_logits' key")
-
-        text_logits = result["text_logits"][:, -1]  # [batch, vocab_size]
-        batch_size = text_logits.shape[0]
-
-        # Use provided generated_tokens or create empty tensor
-        if generated_tokens is None:
-            gen_tokens = torch.empty(batch_size, 0, device=text_logits.device, dtype=torch.long)
-        else:
-            gen_tokens = generated_tokens
-
-        # Apply sampling with top-p and repetition penalty
-        predicted_token = self._sample_text_token(
-            logits=text_logits,
-            generated_tokens=gen_tokens,
-            current_step=current_step,
-            sampling_params=sampling_params,
-        )
-
-        # ASR tokens use greedy decoding (no sampling)
-        asr_predicted_token = result["asr_logits"][:, -1].argmax(dim=-1)
-
-        ans = {
-            "predicted_token": predicted_token,
-            "asr_predicted_token": asr_predicted_token,
-            "cache": result.get("cache", None),
-        }
-        if return_logits:
-            ans["text_logits"] = result["text_logits"]
-            ans["asr_logits"] = result.get("asr_logits")
-            if "function_logits" in result:
-                ans["function_logits"] = result["function_logits"]
-        if "function_logits" in result:
-            ans["function_predicted_token"] = result["function_logits"][:, -1].argmax(dim=-1)
-        return ans
-
-    @staticmethod
-    def _extract_special_token_ids_from_nemo(model) -> set[int]:
-        """
-        Extract special token IDs from NeMo model's tokenizer.
-
-        NeMo tokenizer uses bos_token, eos_token, pad_token (not bos_token_id).
-        Then converts token strings to IDs using token_to_id method.
-
-        Args:
-            model: The DuplexS2SExternalSpeechDecoderModel instance
-
-        Returns:
-            Set of special token IDs, or empty set if extraction fails
-        """
-        special_ids = set()
-        try:
-            tokenizer = model.stt_model.tokenizer
-
-            # Get token strings (NeMo uses bos_token, not bos_token_id)
-            bos_token = getattr(tokenizer, 'bos_token', None)
-            eos_token = getattr(tokenizer, 'eos_token', None)
-            pad_token = getattr(tokenizer, 'pad_token', None)
-
-            # Convert token strings to IDs
-            if hasattr(tokenizer, 'token_to_id'):
-                for token in [bos_token, eos_token, pad_token]:
-                    if token is not None:
-                        tid = tokenizer.token_to_id(token)
-                        if tid is not None and isinstance(tid, int):
-                            special_ids.add(tid)
-        except Exception as e:
-            pass  # Return empty set on failure
-
-        return special_ids
-
-    def to(self, device_or_dtype: torch.device | torch.dtype) -> 'NativeModel':
-        """Move underlying model to device or convert dtype."""
-        self.model = self.model.to(device_or_dtype)
-        return self
-
-    def eval(self) -> 'NativeModel':
-        """Set underlying model to eval mode."""
-        self.model.eval()
-        return self
-
-    @property
-    def device(self) -> torch.device:
-        """Get device of the underlying model."""
-        # Try to get device from model parameters
-        try:
-            return next(self.model.parameters()).device
-        except StopIteration:
-            # No parameters, return CPU
-            return torch.device('cpu')
-
-    def __getattr__(self, name: str):
-        """
-        Delegate attribute access to the underlying model.
-
-        This allows transparent access to model attributes like
-        perception, tokenizer, etc.
-        """
-        # Avoid infinite recursion for special attributes
-        if name in ('model', '__dict__', '__class__'):
-            raise AttributeError(f"'{type(self).__name__}' object has no attribute '{name}'")
-
-        # Delegate to wrapped model
-        return getattr(self.model, name)
-
-
-def create_model(
-    model=None,
-    engine_type: str = "native",
-    vllm_config: dict[str, Any] | None = None,
-    special_token_ids: set[int] | None = None,
-    top_p: float = 1.0,
-    repetition_penalty: float = 1.0,
-    temperature: float = 1.0,
-    **kwargs
-) -> ModelInterface:
-    """
-    Factory function to create appropriate model interface.
-
-    This is the main entry point for creating model interfaces.
-
-    Args:
-        model: The base model to wrap (required for "native" engine, optional for "vllm")
-        engine_type: Type of engine ("native", "vllm")
-        vllm_config: Configuration dict for vLLM engines (required for "vllm")
-        special_token_ids: Set of special token IDs (pad, eos, bos) that should bypass sampling.
-                          If None (default), will auto-extract from model.tokenizer for tokens:
-                          '<s>' (bos), '</s>' (eos), '<SPECIAL_12>' (pad).
-                          You can manually provide: {tokenizer.pad_token_id, tokenizer.eos_token_id, tokenizer.bos_token_id}
-        top_p: Top-p (nucleus) sampling threshold. 1.0 disables it (greedy). Default: 1.0
-        repetition_penalty: Penalty for repeated tokens. 1.0 disables it. Default: 1.0
-        temperature: Temperature for sampling. 1.0 = no change, 0.0 = greedy. Default: 1.0
-        **kwargs: Additional arguments passed to the interface constructor
-
-    Returns:
-        ModelInterface instance
-
-    Example:
-        >>> # Use native PyTorch model with greedy decoding (default)
-        >>> interface = create_model(model, engine_type="native")
-        >>>
-        >>> # Use native with top-p sampling (special_token_ids auto-extracted from model.tokenizer)
-        >>> # Auto-extracts IDs for: '<s>', '</s>', '<SPECIAL_12>'
-        >>> interface = create_model(
-        >>>     model,
-        >>>     engine_type="native",
-        >>>     top_p=0.9
-        >>> )
-        >>>
-        >>> # Use native with top-p and repetition penalty (auto-extract special tokens)
-        >>> interface = create_model(
-        >>>     model,
-        >>>     engine_type="native",
-        >>>     top_p=0.9,
-        >>>     repetition_penalty=1.2
-        >>> )
-        >>>
-        >>> # Manually provide special_token_ids (if auto-extraction fails or you want custom tokens)
-        >>> special_ids = {
-        >>>     tokenizer.pad_token_id,
-        >>>     tokenizer.eos_token_id,
-        >>>     tokenizer.bos_token_id
-        >>> }
-        >>> interface = create_model(
-        >>>     model,
-        >>>     engine_type="native",
-        >>>     special_token_ids=special_ids,
-        >>>     top_p=0.9,
-        >>>     repetition_penalty=1.2
-        >>> )
-        >>>
-        >>> # Use vLLM with streaming engine
-        >>> vllm_cfg = {
-        >>>     "model_path": "/path/to/vllm/checkpoint",
-        >>>     "max_model_len": 10240,
-        >>>     "gpu_memory_utilization": 0.8,
-        >>>     "dtype": "bfloat16"
-        >>> }
-        >>> interface = create_model(
-        >>>     engine_type="vllm",
-        >>>     vllm_config=vllm_cfg
-        >>> )
-        >>>
-        >>> # Perform inference
-        >>> result = interface(input_embeds, cache=cache)
-        >>>
-        >>> # For repetition penalty, pass generated_tokens and current_step
-        >>> result = interface(input_embeds, cache=cache, generated_tokens=prev_tokens, current_step=step)
-    """
-    engine_type = engine_type.lower()
-
-    if engine_type == "native":
-        if model is None:
-            raise ValueError("model must be provided for native engine")
-        return NativeModel(
-            model=model,
-            special_token_ids=special_token_ids,
-            top_p=top_p,
-            repetition_penalty=repetition_penalty,
-            temperature=temperature,
-        )
-
-    elif engine_type == "vllm_eartts":
-        if vllm_config is None:
-            raise ValueError("vllm_config must be provided for vLLM EARTTS engine")
-        # VllmEARTTSModel for TTS inference
-        return VllmEARTTSModel(
-            **vllm_config,
-            model_type="eartts",
-            special_token_ids=special_token_ids,
-            top_p=top_p,
-            repetition_penalty=repetition_penalty,
-            temperature=temperature,
-            **kwargs
-        )
-
-    elif engine_type.startswith("vllm"):
-        if vllm_config is None:
-            raise ValueError("vllm_config must be provided for vLLM engine")
-        # VllmLLMModel doesn't need the PyTorch model, only the config
-        return VllmLLMModel(
-            **vllm_config,
-            model_type="llm",
-            special_token_ids=special_token_ids,
-            top_p=top_p,
-            repetition_penalty=repetition_penalty,
-            temperature=temperature,
-            **kwargs
-        )
-
-    else:
-        raise ValueError(
-            f"Unknown engine_type: {engine_type}. "
-            f"Supported types: 'native', 'vllm', 'vllm_llm', 'vllm_eartts'"
-        )
